requirements.txt
# Pin with: pip freeze after install
Django>=5.0,<6
djangorestframework>=3.15
drf-spectacular>=0.27
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
django-environ>=0.11
Pillow>=10.0
reportlab>=4.0
